Remove the duplicated runtime parse helpers
