+++ src/Meziantou.Framework.TemporaryContainers/ContainerRuntimes/ExecutableContainerRuntime.cs
@@ -388,16 +388,4 @@ internal abstract class ExecutableContainerRuntime : ContainerRuntime
     {
         return value.Replace("'", "''", StringComparison.Ordinal);
     }
-
-    protected static DateTimeOffset? ParseDate(string? value)
-    {
-        if (string.IsNullOrEmpty(value))
-            return null;
-
-        if (!DateTimeOffset.TryParse(value, CultureInfo.InvariantCulture, DateTimeStyles.RoundtripKind, out var result))
-            return null;
-
-        // Runtimes report the zero date for events that never happened (for example FinishedAt on a running container).
-        return result.UtcDateTime.Year <= 1 ? null : result;
-    }
 }
